+++ lib/add_prodact/add_prodact.dart
@@ -1,0 +1,1 @@
+export 'view/view_page.dart';
