+++ lib/add_prodact/view_prodact/view_prodact.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ViewProdact extends StatefulWidget {
+  const ViewProdact({super.key});
+
+  @override
+  State<ViewProdact> createState() => _ViewProdactState();
+}
+
+class _ViewProdactState extends State<ViewProdact> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+    );
+  }
+}
