+++ lib/app/app.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+import '../view/view_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const ViewPage(),
+    );
+  }
+}
