+++ lib/view/view_page.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ViewPage extends StatefulWidget {
+  const ViewPage({super.key});
+
+  @override
+  State<ViewPage> createState() => _ViewPageState();
+}
+
+class _ViewPageState extends State<ViewPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('jarnama app'),),
+    );
+  }
+}
